+++ data/style.css~
@@ -1,0 +1,73 @@
+@import url(http://fonts.googleapis.com/css?family=Open+Sans);
+body {
+	background: url(blackprint.png);
+	background-repeat: repeat;
+	font-family: "Open Sans";
+	color rgb(250, 250, 250);
+}
+
+#linklogo {
+	position: absolute;
+	top: 7px;
+	left: 9px;
+}
+#mbtn {
+	position: absolute;
+	top: 18px;
+	right: 19px;
+}
+#solgan {
+	position: absolute;
+	top: 17px;
+	left: 138px;
+	color: rgb(250, 250, 250);
+}
+
+#main {
+	margin-top: 51px;
+}
+#buttons {
+	float: right;
+}
+	#buttons button {
+		padding: 2px 4px;
+		font-size: 12px;
+		font-weight: bold;
+		line-height: 14px;
+		border: none;
+		border-radius: 3px;
+		color: rgb(250, 250, 250);
+	}	
+		#hoerer { background: rgb(58, 135, 173); }
+		#google { background: rgb(66, 133, 244); }
+		#jamendo { background: rgb(105, 62, 105); }
+		#bandcamp { background: rgb(97, 154, 169); }
+#knopf {
+	margin-left: 7px;
+	margin-top: 25px;
+}
+#menu {
+	color: #fafafa;
+	padding-left: 25px;
+}
+#status {
+	position: fixed;
+	left: 0;
+	bottom: 0;
+	width: 100%;
+	padding: 3px;
+	padding-left: 12px;
+	color: #fafafa;
+	box-shadow: 0px 2px 5px 4px #0072be;
+}
+
+.hide { display: none; }
+
+a {
+	text-decoration: none;
+	color: #fff;
+}
+a:hover {
+	text-decoration: underline;
+	color: #fff;
+}
