+++ package.json
@@ -1,0 +1,10 @@
+{
+  "name": "theradiocc-player",
+  "title": "TheRadio.CC - Player",
+  "fullName": "TheRadio.CC - Player",
+  "id": "jid0-BlOtEktFWTPtfNiQElNtI3ANfcA",
+  "description": "Listen to TheRadio.CC in your Firefox Webbrowser",
+  "author": "ventos",
+  "license": "GPL 2.0",
+  "version": "0.3.0.0"
+}
