--- src/data/player.html
@@ -1,47 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-    <head>
-        <title>TheRadio.CC - Player</title>
-		<link rel="stylesheet" type="text/css" href="style.css" />
-        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-		<script type="text/javascript" src="jquery-2.0.0.min.js"></script>
-		<script type="text/javascript" src="script.js"></script>
-	</head>
-	<body>
-		<audio id="audioplayer" preload="none">
-			<source src="http://ogg.theradio.cc" type="audio/ogg" /> 
-		</audio>
-		<div id="header">
-			<table>
-				<tr>
-					<td><a href="http://theradio.cc" target="_blank"><img src="logo.png" id="logo" width="120" height="42" /></a></td>
-					<td>freie Musik f&uuml;r freie H&ouml;rer&nbsp;&nbsp;</td>
-					<td><a href="javascript:menu()"><img src="menu.png"></a></td>
-				</tr>
-			</table>
-		</div>
-		<div id="main">
-			<div id="start">
-				<br>
-				<div id="suche">
-					<button id="hoerer"><span id="listeners">0</span> H&ouml;rer</button><br>
-					<button id="google" onclick="search('google')">Google</button><br>
-					<button id="jamendo" onclick="search('jamendo')">Jamendo</button><br>
-					<button id="bandcamp" onclick="search('bandcamp')">Bandcamp</button>
-				</div>
-				<div>
-					<img src="play.png" width="50" height="50" onclick="play()" id="knopf"/>
-				</div>
-			</div>
-			<div id="menu" class="hide">
-				<strong>Infos</strong><br>
-				Der Stream ist von <a href="http://theradio.cc/" target="_blank">TheRadio.CC</a> und
-				das Addon von <a href="https://github.com/ventos/" target="_blank">ventos</a>.<br>
-				<a href="update.html" target="_blank">Update</a>
-			</div>
-			<div id="controls">
-				<span id="nowPlaying"><em>Please check your Internet connection</em></span>
-			</div>
-		</div>
-	</body>
-</html>
