--- src/data/style.css
@@ -1,113 +1,0 @@
-/* Allgemeines Zeugs */
-@import url(http://fonts.googleapis.com/css?family=Open+Sans);
-
-body {
-    font-family: 'Open Sans';
-    background: #292929;
-    background-repeat:repeat;
-	background-attachment:fixed;
-	color: rgb(250, 250, 250);
-}
-
-.hide    	{ display: none;	}
-
-#start {
-	margin-left: 1em;
-}
-
-#knopf {
-	float: left;    
-}
-
-#nowPlaying {
-	float: left;
-	padding-left: 10px;
-	padding-top: 4px;
-	padding-bottom: 4px;
-}
-
-#menu {
-	padding-left: 10px;
-}
-
-#info {
-	padding-left: 10px;    
-}
-
-#hoerer {
-	padding: 2px 4px;
-	font-size: 12px;;
-	font-weight: bold;
-	line-height: 14px;
-	color: rgb(250, 250, 250);
-	border-radius: 3px;
-	background: rgb(58, 135, 173);
-	padding: 2px 4px;
-	border: none;
-}
-#suche {
-	display: inline-block;
-	float: right;
-	margin-right: 33px;
-}
-
-#google {
-	background: rgb(66, 133, 244);
-	padding: 2px 4px;
-	font-size: 12px;
-	font-weight: bold;
-	line-height: 14px;
-	color: rgb(250, 250, 250);
-	border-radius: 3px;
-	border: none;
-}
-#jamendo {
-	background: rgb(105, 62, 105);
-	padding: 2px 4px;
-	font-size: 12px;
-	font-weight: bold;
-	line-height: 14px;
-	color: rgb(250, 250, 250);
-	border-radius: 3px;
-	border: none;
-}
-#bandcamp {
-	background: rgb(97, 154, 169);
-	padding: 2px 4px;
-	font-size: 12px;
-	font-weight: bold;
-	line-height: 14px;
-	color: rgb(250, 250, 250);
-	border-radius: 3px;
-	border: none;
-}
-
-#logo {
-	padding-top: 3px;
-}
-
-a {
-	text-decoration: none;
-	color: rgb(250, 250, 250);
-}
-
-a:hover {
-	text-decoration: underline;
-	color: rgb(250, 250, 250);
-}
-
-#controls {
-	background: #252525;
-	position: fixed;
-    bottom: 0px;
-	left: 0px;
-	width: 100%;
-	padding: 3px;
-	color: rgb(250, 250, 250);
-	box-shadow: 0px 2px 5px 4px #e9e9e9; ;
-}
-
-::-moz-selection {
-	background-color: #444; 
-	color: rgb(250, 250, 250);
-}
