--- src/data/update.css
@@ -1,29 +1,0 @@
-@import url(http://fonts.googleapis.com/css?family=Open+Sans:300,400);
-
-body {
-    font-family: 'Open Sans';
-    background: rgb(244, 244, 244);
-	color: #0a0a0a;
-}
-
-#header {
-    background: #202530;
-	height: 4em;
-	width: 100%;
-	top: 0px;
-	left: 0px;
-	position: fixed;
-	box-shadow: 0px 0px 4px 4px rgb(0, 0, 0);
-}
-h1.heading {
-    margin-left: 3%;
-	text-align: left;
-	font-size: 2em;
-	font-family: 'Open Sans';
-	font-weight: 300;
-	color: rgb(250, 250, 250);
-}
-
-#content {
-    
-}
