--- src/data/update.html
@@ -1,17 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-    <head>
-		<title>Updates</title>
-		<link rel="stylesheet" type="text/css" href="update.css" />
-        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-		<script type="text/javascript" src="jquery-2.0.0.min.js"></script>
-	</head>
-	<body>
-        <div id="header">
-    		<h1 class="heading">Updates</h1>
-		</div>
-        <div id="content">
-		    <a href="https://raw.github.com/ventos/trccplayerforfirefox/master/theradiocc-player.xpi">Download</a>
-        </div>
-	</body>
-</html>
