--- src/package.json
@@ -1,1 +1,0 @@
-{"license": "", "description": "Listen to TheRadio.CC in your Firefox Webbrowser", "contributors": [], "author": "ventos", "url": "", "dependencies": ["api-utils", "addon-kit"], "version": "0.2.0.6.rev59", "lib": "lib", "fullName": "TheRadio.CC - Player", "main": "main", "id": "jid0-BlOtEktFWTPtfNiQElNtI3ANfcA", "name": "theradiocc-player"}
